test(EntityBinder): 테스트의 위치를 올바르게 변경
