+++ src/test/java/persistence/sql/model/ColumnTest.java
@@ -55,67 +55,4 @@ class ColumnTest {
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("This field is not a column: index");
     }
-
-    @DisplayName("올바르게 인스턴스의 값을 반환하는지 확인")
-    @Test
-    void getValue() throws NoSuchFieldException {
-        Column column = createPerson3NameColumn();
-        Person3 person = new Person3("a", 1, "email@email.com");
-
-        EntityBinder entityBinder = new EntityBinder(person);
-        Object result = entityBinder.getValue(column);
-
-        assertThat(result).isEqualTo("a");
-    }
-
-    @DisplayName("해당 필드가 없는 instance를 넣었을 경우 IllegalArgumentException을 던진다.")
-    @Test
-    void getValueWithException() throws NoSuchFieldException {
-//        Column column = createPerson3NameColumn();
-//        Person2 person = new Person2("a", 1, "email@email.com");
-//
-//        assertThatThrownBy(() -> column.getValue(person))
-//                .isInstanceOf(IllegalArgumentException.class)
-//                .hasMessageContaining("This instance does not have any of the fields in that column.");
-    }
-
-    @DisplayName("인스턴스의 필드에 값을 정상적으로 할당하는지 확인")
-    @Test
-    void setValue() throws Exception {
-        Person3 person = new Person3("qweqwe", 123, "e");
-        Column column = createPerson3NameColumn();
-
-//        column.setValue(person, "aaa");
-//
-//        String result = person.getName();
-//        assertThat(result).isEqualTo("aaa");
-    }
-
-    @DisplayName("필드가 없는 instance에 값을 넣으려는 경우 IllegalArgumentException을 던진다.")
-    @Test
-    void setValueWithWrongInstance() throws Exception {
-//        Person2 person = new Person2("a", 1, "email@email.com");
-//        Column column = createPerson3NameColumn();
-//
-//        assertThatThrownBy(() -> column.setValue(person, "some name"))
-//                .isInstanceOf(IllegalArgumentException.class)
-//                .hasMessageContaining("Invalid instance or value: ");
-    }
-
-    @DisplayName("필드에 잘못된 값을 넣으려는 경우 IllegalArgumentException을 던진다.")
-    @Test
-    void setValueWithWrongType() throws Exception {
-//        Person3 person = new Person3("aa", 2, "email@email.com");
-//        Column column = createPerson3NameColumn();
-//
-//        assertThatThrownBy(() -> column.setValue(person, 112323))
-//                .isInstanceOf(IllegalArgumentException.class)
-//                .hasMessageContaining("Invalid instance or value: ");
-    }
-
-    private Column createPerson3NameColumn() throws NoSuchFieldException {
-        Class<Person3> clazz = Person3.class;
-        Field field = clazz.getDeclaredField("name");
-        return new Column(field);
-    }
 }
